added feature flag for analytics dashboard
